Initial cleanup of association decorators
