+++ src/app/models/AudioRecording.ts
@@ -3,13 +3,8 @@ import { ACCOUNT, SHALLOW_SITE } from "@baw-api/ServiceTokens";
 import { Duration } from "luxon";
 import { Observable } from "rxjs";
 import { DateTimeTimezone, Id, Uuid } from "../interfaces/apiInterfaces";
-import {
-  AbstractModel,
-  Creator,
-  Deleter,
-  HasOne,
-  Updater,
-} from "./AbstractModel";
+import { AbstractModel } from "./AbstractModel";
+import { Creator, Deleter, HasOne, Updater } from "./AssociationDecorators";
 import { BawDateTime, BawDuration } from "./AttributeDecorators";
 import type { Site } from "./Site";
 import type { User } from "./User";
@@ -83,9 +78,9 @@ export class AudioRecording extends AbstractModel implements IAudioRecording {
   public updater?: Observable<User>;
   @Deleter<AudioRecording>()
   public deleter?: Observable<User>;
-  @HasOne(ACCOUNT, (m: AudioRecording) => m.uploaderId)
+  @HasOne<AudioRecording>(ACCOUNT, "uploaderId")
   public uploader?: Observable<User>;
-  @HasOne(SHALLOW_SITE, (m: AudioRecording) => m.siteId)
+  @HasOne<AudioRecording>(SHALLOW_SITE, "siteId")
   public site?: Observable<Site>;
 
   constructor(audioRecording: IAudioRecording, injector?: Injector) {
