+++ src/app/models/Dataset.ts
@@ -6,7 +6,8 @@ import {
   Param,
 } from "@interfaces/apiInterfaces";
 import { Observable } from "rxjs";
-import { AbstractModel, Creator, Updater } from "./AbstractModel";
+import { AbstractModel } from "./AbstractModel";
+import { Creator, Updater } from "./AssociationDecorators";
 import { BawDateTime, BawPersistAttr } from "./AttributeDecorators";
 import type { User } from "./User";
