+++ src/app/models/Script.ts
@@ -2,7 +2,8 @@ import { Injector } from "@angular/core";
 import { adminScriptMenuItem } from "@component/admin/scripts/scripts.menus";
 import { Observable } from "rxjs";
 import { DateTimeTimezone, Id, Param } from "../interfaces/apiInterfaces";
-import { AbstractModel, Creator } from "./AbstractModel";
+import { AbstractModel } from "./AbstractModel";
+import { Creator } from "./AssociationDecorators";
 import { BawDateTime, BawPersistAttr } from "./AttributeDecorators";
 import type { User } from "./User";
